refactor(ts): move patch protocol types to shared types
Move PATCH protocol types, Component/Runtime, and related patch operation types into ts/miso/types.ts to centralize type exports. Update patch application, patch context generator, test imports, and barrel exports to reference types.ts while keeping patch() implementation in patch.ts.
