+++ js/miso.prod.js
@@ -1,1 +1,1 @@
-var F="1.9.0.0";function L(U,z){var G=function(){var K=document.getElementById(U);if(K&&K.focus)K.focus()};z>0?setTimeout(G,z):G()}function I(U,z){var G=function(){var K=document.getElementById(U);if(K&&K.blur)K.blur()};z>0?setTimeout(G,z):G()}function w(U,z){var G=function(){var K=document.getElementById(U);if(K&&typeof K.select==="function")K.select()};z>0?setTimeout(G,z):G()}function V(U,z,G,K){var H=function(){var X=document.getElementById(U);if(X&&typeof X.setSelectionRange==="function")X.setSelectionRange(z,G,"none")};K>0?setTimeout(H,K):H()}function O(U,z,G,K,H,X,Q){var Z={method:z,headers:K};if(G)Z.body=G;let W={},$=null;try{fetch(U,Z).then((Y)=>{$=Y.status;for(let[_,J]of Y.headers)W[_]=J;if(!Y.ok)throw new Error(Y.statusText);if(Q=="json")return Y.json();else if(Q=="text")return Y.text();else if(Q==="arrayBuffer")return Y.arrayBuffer();else if(Q==="blob")return Y.blob();else if(Q==="bytes")return Y.bytes();else if(Q==="formData")return Y.formData();else if(Q==="none")return H({error:null,body:null,headers:W,status:$})}).then((Y)=>H({error:null,body:Y,headers:W,status:$})).catch((Y)=>X({error:null,body:Y,headers:W,status:$}))}catch(Y){X({body:null,error:Y.message,headers:W,status:$})}}function P(U,z,G,K,H,X,Q,Z,W){try{let $=new WebSocket(U);return $.onopen=function(){z()},$.onclose=function(Y){G(Y)},$.onerror=function(Y){console.error(Y),Z("WebSocket error received")},$.onmessage=function(Y){if(typeof Y.data==="string")try{if(W){if(K)K(Y.data);return}let _=JSON.parse(Y.data);if(H)H(_)}catch(_){if(W&&K)K(Y.data);else Z(_.message)}else if(Y.data instanceof Blob){if(X)X(Y.data)}else if(Y.data instanceof ArrayBuffer){if(Q)Q(Y.data)}else console.error("Received unknown message type from WebSocket",Y),Z("Unknown message received from WebSocket")},$}catch($){Z($.message)}}function f(U){if(U)U.close(),U=null}function j(U,z){if(z&&U&&U.readyState===WebSocket.OPEN)U.send(z)}function b(U,z,G,K,H,X){try{let Q=new EventSource(U);return Q.onopen=function(){z()},Q.onerror=function(){H("EventSource error received")},Q.onmessage=function(Z){try{if(X){if(G)G(Z.data);return}let W=JSON.parse(Z.data);if(K)K(W)}catch(W){if(X&&G)G(Z.data);else H(W.message)}},Q}catch(Q){H(Q.message)}}function T(U){if(U)U.close(),U=null}function k(U,z){if(!U.classList)U.classList=new Set;for(let G of z)for(let K of G.trim().split(" "))if(K)U.classList.add(K)}function h(U,z){if(!U.parent)return;z.nextSibling=U.nextSibling?null:U.nextSibling,z.parent=U.parent,U.parent.child=z}function m(U,z={}){let G=Object.keys(z),K=Object.values(z);return new Function(...G,U)(...K)}function g(U){if(U===null||U===void 0)return 0;if(typeof U==="number")return 1;if(typeof U==="string")return 2;if(typeof U==="boolean")return 3;if(Array.isArray(U))return 4;return 5}function y(U){return function(){U|=0,U=U+2654435769|0;var z=U^U>>>15;return z=Math.imul(z,2246822507),z=z^z>>>13,z=Math.imul(z,3266489909),((z^z>>>16)>>>0)/4294967296}}function u(){let U=new Uint32Array(1);return crypto.getRandomValues(U)[0]}function v(){return Math.random()}function q(U){switch(U.type){case 0:return S(U);default:return U.domRef}}function S(U){if(!U.child)throw new Error("'drill' called on an unmounted Component. This should never happen, please make an issue.");switch(U.child.type){case 0:return S(U.child);default:return U.child.domRef}}function A(U,z,G,K){if(!U&&!z)return;else if(!U)BU(z,G,K);else if(!z)c(U,G,K);else if(U.type===2&&z.type===2)XU(U,z,K);else if(U.type===0&&z.type===0){if(z.key===U.key){if(z.child=U.child,z.componentId=U.componentId,U.child)U.child.parent=z;return}l(U,z,G,K)}else if(U.type===1&&z.type===1)if(z.tag===U.tag&&z.key===U.key)z.domRef=U.domRef,o(U,z,K);else l(U,z,G,K);else l(U,z,G,K)}function XU(U,z,G){if(U.text!==z.text)G.setTextContent(U.domRef,z.text);z.domRef=U.domRef;return}function l(U,z,G,K){switch(U.type){case 2:break;default:M(U);break}switch(i(G,1,q(U),z,K),U.type){case 2:break;default:R(U);break}}function c(U,z,G){switch(U.type){case 2:break;default:M(U);break}switch(G.removeChild(z,q(U)),U.type){case 2:break;default:R(U);break}}function R(U){switch(YU(U),U.type){case 1:for(let z of U.children)if(z.type===1||z.type===0)R(z);break;case 0:if(U.child){if(U.child.type===1||U.child.type===0)R(U.child)}break}}function YU(U){if(U.type===1&&U.onDestroyed)U.onDestroyed();if(U.type===0)DU(U)}function ZU(U){switch(U.type){case 0:break;case 1:if(U.onBeforeDestroyed)U.onBeforeDestroyed();break;default:break}}function M(U){switch(ZU(U),U.type){case 1:for(let z of U.children){if(z.type===2)continue;M(z)}break;case 0:if(U.child){if(U.child.type===1||U.child.type===0)M(U.child)}break}}function o(U,z,G){WU(U?U.props:{},z.props,z.domRef,z.ns==="svg",G),$U(U?U.classList:null,z.classList,z.domRef,G),_U(U?U.css:{},z.css,z.domRef,G),AU(U?U.children:[],z.children,z.domRef,G),EU(z)}function $U(U,z,G,K){if(!U&&!z)return;if(!U){for(let H of z)K.addClass(H,G);return}if(!z){for(let H of U)K.removeClass(H,G);return}for(let H of U)if(!z.has(H))K.removeClass(H,G);for(let H of z)if(!U.has(H))K.addClass(H,G);return}function WU(U,z,G,K,H){var X;for(let Q in U)if(X=z[Q],X===void 0)if(K||!(Q in G)||Q==="disabled")H.removeAttribute(G,Q);else H.setAttribute(G,Q,"");else{if(X===U[Q]&&Q!=="checked"&&Q!=="value")continue;if(K)if(Q==="href")H.setAttributeNS(G,"http://www.w3.org/1999/xlink","href",X);else H.setAttribute(G,Q,X);else if(Q in G&&!(Q==="list"||Q==="form"))G[Q]=X;else H.setAttribute(G,Q,X)}for(let Q in z){if(U&&U[Q])continue;if(X=z[Q],K)if(Q==="href")H.setAttributeNS(G,"http://www.w3.org/1999/xlink","href",X);else H.setAttribute(G,Q,X);else if(Q in G&&!(Q==="list"||Q==="form"))G[Q]=z[Q];else H.setAttribute(G,Q,X)}}function _U(U,z,G,K){K.setInlineStyle(U,z,G)}function qU(U,z){if(U.length===0||z.length===0)return!1;for(var G=0;G<U.length;G++)if(U[G].key===null||U[G].key===void 0)return!1;for(var G=0;G<z.length;G++)if(z[G].key===null||z[G].key===void 0)return!1;return!0}function AU(U,z,G,K){if(qU(U,z))RU(U,z,G,K);else for(let H=0;H<Math.max(z.length,U.length);H++)A(U[H],z[H],G,K)}function JU(U,z){if(U.ns==="svg")U.domRef=z.createElementNS("http://www.w3.org/2000/svg",U.tag);else if(U.ns==="mathml")U.domRef=z.createElementNS("http://www.w3.org/1998/Math/MathML",U.tag);else U.domRef=z.createElement(U.tag)}function n(U,z,G){if(z.onCreated)z.onCreated(z.domRef)}function i(U,z,G,K,H){switch(K.type){case 2:switch(K.domRef=H.createTextNode(K.text),z){case 2:H.insertBefore(U,K.domRef,G);break;case 0:H.appendChild(U,K.domRef);break;case 1:H.replaceChild(U,K.domRef,G);break}break;case 0:SU(U,z,G,K,H);break;case 1:if(K.onBeforeCreated)K.onBeforeCreated();if(JU(K,H),K.onCreated)K.onCreated(K.domRef);switch(o(null,K,H),z){case 2:H.insertBefore(U,K.domRef,G);break;case 0:H.appendChild(U,K.domRef);break;case 1:H.replaceChild(U,K.domRef,G);break}break}}function EU(U){if(U.tag==="canvas"&&U.draw)U.draw(U.domRef)}function DU(U){U.unmount(U.componentId)}function SU(U,z,G,K,H){let X=K.mount(U);if(K.componentId=X.componentId,K.child=X.componentTree,X.componentTree.parent=K,X.componentTree.type!==0){let Q=q(X.componentTree);if(z===1&&G)H.replaceChild(U,Q,G);else if(z===2)H.insertBefore(U,Q,G)}}function BU(U,z,G){i(z,0,null,U,G)}function N(U,z,G,K){K.insertBefore(U,q(z),G?q(G):null)}function NU(U,z,G,K){K.swapDOMRefs(q(U),q(z),G)}function RU(U,z,G,K){var H=0,X=0,Q=U.length-1,Z=z.length-1,W,$,Y,_,J,C,r;for(;;){if(X>Z&&H>Q)break;if($=z[X],Y=z[Z],J=U[H],_=U[Q],H>Q)A(null,$,G,K),N(G,$,J,K),U.splice(X,0,$),X++;else if(X>Z){W=Q;while(Q>=H)c(U[Q--],G,K);U.splice(H,W-H+1);break}else if(J.key===$.key)A(U[H++],z[X++],G,K);else if(_.key===Y.key)A(U[Q--],z[Z--],G,K);else if(J.key===Y.key&&$.key===_.key)NU(_,J,G,K),MU(U,H,Q),A(U[H++],z[X++],G,K),A(U[Q--],z[Z--],G,K);else if(J.key===Y.key)N(G,J,_.nextSibling,K),U.splice(Q,0,U.splice(H,1)[0]),A(U[Q--],z[Z--],G,K);else if(_.key===$.key)N(G,_,J,K),U.splice(H,0,U.splice(Q,1)[0]),A(U[H++],$,G,K),X++;else{C=!1,W=H;while(W<=Q){if(U[W].key===$.key){C=!0,r=U[W];break}W++}if(C)U.splice(H,0,U.splice(W,1)[0]),A(U[H++],$,G,K),N(G,r,U[H],K),X++;else i(G,2,q(J),$,K),U.splice(H++,0,$),X++,Q++}}}function MU(U,z,G){let K=U[z];U[z]=U[G],U[G]=K}function UU(U,z,G,K,H){for(let X of z)H.addEventListener(U,X.name,function(Q){CU(Q,U,G,K,H)},X.capture)}function CU(U,z,G,K,H){G(function(X){if(Array.isArray(U))for(let Q of U)t(Q,X,z,K,H);else t(U,X,z,K,H)})}function t(U,z,G,K,H){var X=H.getTarget(U);if(X){let Q=FU(G,X,H);D(U,z,Q,K,H)}}function FU(U,z,G){var K=[];while(!G.isEqual(U,z))if(K.unshift(z),z&&G.parentNode(z))z=G.parentNode(z);else return K;return K}function D(U,z,G,K,H){if(!G.length){if(K)console.warn('Event "'+U.type+'" did not find an event handler to dispatch on',z,U);return}else if(G.length>1){if(z.type===2)return;else if(z.type===0){if(!z.child){if(K)throw console.error("VComp has no child property set during event delegation",z),console.error("This means the Component has not been fully mounted, this should never happen"),new Error("VComp has no .child property set during event delegation");return}return D(U,z.child,G,K,H)}else if(z.type===1){if(H.isEqual(z.domRef,G[0])){let X=z.events.captures[U.type];if(X){let Q=X.options;if(Q.preventDefault)U.preventDefault();if(!U.captureStopped)X.runEvent(U,z.domRef);if(Q.stopPropagation)U.captureStopped=!0}G.splice(0,1);for(let Q of z.children)if(H.isEqual(q(Q),G[0]))D(U,Q,G,K,H)}return}}else if(z.type===0){if(z.child)D(U,z.child,G,K,H)}else if(z.type===1){let X=z.events.captures[U.type];if(X&&!U.captureStopped){let Z=X.options;if(H.isEqual(G[0],z.domRef)){if(Z.preventDefault)U.preventDefault();if(X.runEvent(U,G[0]),Z.stopPropagation)U.captureStopped=!0}}let Q=z.events.bubbles[U.type];if(Q&&!U.captureStopped){let Z=Q.options;if(H.isEqual(G[0],z.domRef)){if(Z.preventDefault)U.preventDefault();if(Q.runEvent(U,G[0]),!Z.stopPropagation)e(z.parent,U)}}else if(!U.captureStopped)e(z.parent,U)}}function e(U,z){while(U)switch(U.type){case 2:break;case 1:let G=U.events.bubbles[z.type];if(G){let K=G.options;if(K.preventDefault)z.preventDefault();if(G.runEvent(z,U.domRef),K.stopPropagation)return}U=U.parent;break;case 0:if(!U.eventPropagation)return;U=U.parent;break}}function B(U,z){if(typeof U[0]==="object"){var G=[];for(var K=0;K<U.length;K++)G.push(B(U[K],z));return G}for(let Z of U)z=z[Z];var H;if(z instanceof Array||"length"in z&&z.localName!=="select"){H=[];for(var X=0;X<z.length;X++)H.push(B([],z[X]));return H}H={};for(var Q in LU(z)){if(z.localName==="input"&&(Q==="selectionDirection"||Q==="selectionStart"||Q==="selectionEnd"))continue;if(typeof z[Q]=="string"||typeof z[Q]=="number"||typeof z[Q]=="boolean")H[Q]=z[Q]}return H}function LU(U){var z={},G=0;do{var K=Object.getOwnPropertyNames(U);for(G=0;G<K.length;G++)z[K[G]]=null}while(U=Object.getPrototypeOf(U));return z}function IU(U){var z=0,G=U.length>0?[U[0]]:[];for(var K=1;K<U.length;K++){if(G[z].type===2&&U[K].type===2){G[z].text+=U[K].text;continue}G[++z]=U[K]}return G}function d(U,z,G,K,H){if(!G||!z)return!1;if(z.nodeType===3)return!1;if(!p(U,G,K.firstChild(z),K,H)){if(U)console.warn("[DEBUG_HYDRATE] Could not copy DOM into virtual DOM, falling back to diff");while(K.firstChild(z))H.removeChild(z,K.lastChild(z));return!1}else if(U)console.info("[DEBUG_HYDRATE] Successfully prerendered page");return!0}function x(U,z,G){if(U)console.warn("[DEBUG_HYDRATE] VTree differed from node",z,G)}function zU(U){if(U.substr(0,1)=="#"){let z=(U.length-1)/3,G=[17,1,0.062272][z-1];return[Math.round(parseInt(U.substr(1,z),16)*G),Math.round(parseInt(U.substr(1+z,z),16)*G),Math.round(parseInt(U.substr(1+2*z,z),16)*G)]}else return U.split("(")[1].split(")")[0].split(",").map((z)=>{return+z})}function a(U,z,G){return GU(!0,U,z,G)}function GU(U,z,G,K){if(z.type==2){if(G.getTag(z.domRef)!=="#text")console.warn("VText domRef not a TEXT_NODE",z),U=!1;else if(z.text!==G.getTextContent(z.domRef))console.warn("VText node content differs",z),U=!1}else if(z.type===1){if(z.tag.toUpperCase()!==G.getTag(z.domRef).toUpperCase())console.warn("Integrity check failed, tags differ",z.tag.toUpperCase(),G.getTag(z.domRef)),U=!1;if("children"in z&&z.children.length!==G.children(z.domRef).length)console.warn("Integrity check failed, children lengths differ",z,z.children,G.children(z.domRef)),U=!1;for(let H in z.props)if(H==="href"||H==="src"){let X=window.location.origin+"/"+z.props[H],Q=G.getAttribute(z.domRef,H),Z=z.props[H];if(X!==Q&&Z!==Q&&Z+"/"!==Q&&X+"/"!==Q)console.warn("Property "+H+" differs",z.props[H],G.getAttribute(z.domRef,H)),U=!1}else if(H==="height"||H==="width"){if(parseFloat(z.props[H])!==parseFloat(G.getAttribute(z.domRef,H)))console.warn("Property "+H+" differs",z.props[H],G.getAttribute(z.domRef,H)),U=!1}else if(H==="class"||H==="className"){if(z.props[H]!==G.getAttribute(z.domRef,"class"))console.warn("Property class differs",z.props[H],G.getAttribute(z.domRef,"class")),U=!1}else if(z.props[H]!==G.getAttribute(z.domRef,H))console.warn("Property "+H+" differs",z.props[H],G.getAttribute(z.domRef,H)),U=!1;for(let H in z.css)if(H==="color"){if(zU(G.getInlineStyle(z.domRef,H)).toString()!==zU(z.css[H]).toString())console.warn("Style "+H+" differs",z.css[H],G.getInlineStyle(z.domRef,H)),U=!1}else if(z.css[H]!==G.getInlineStyle(z.domRef,H))console.warn("Style "+H+" differs",z.css[H],G.getInlineStyle(z.domRef,H)),U=!1;for(let H of z.children){let X=GU(U,H,G,K);U=U&&X}}return U}function p(U,z,G,K,H){switch(z.type){case 0:if(!p(U,z.child,G,K,H))return!1;break;case 2:if(G.nodeType!==3||z.text.trim()!==G.textContent.trim())return x(U,z,G),!1;z.domRef=G;break;case 1:if(G.nodeType!==1)return x(U,z,G),!1;z.domRef=G,z.children=IU(z.children),n(G,z,H);for(var X=0;X<z.children.length;X++){let Q=z.children[X],Z=G.childNodes[X];if(!Z)return x(U,Q,Z),!1;if(!p(U,Q,Z,K,H))return!1}break}return!0}var s={addEventListener:(U,z,G,K)=>{U.addEventListener(z,G,K)},delegator:(U,z,G,K,H)=>{UU(U,z,G,K,H)},isEqual:(U,z)=>{return U===z},getTarget:(U)=>{return U.target},parentNode:(U)=>{return U.parentNode}},HU={getInlineStyle:(U,z)=>{return U.style[z]},firstChild:(U)=>{return U.firstChild},lastChild:(U)=>{return U.lastChild},getAttribute:(U,z)=>{if(z==="class")return U.className;if(z in U)return U[z];return U.getAttribute(z)},getTag:(U)=>{return U.nodeName},getTextContent:(U)=>{return U.textContent},children:(U)=>{return U.childNodes}},KU={mountComponent:function(U,z,G){return},unmountComponent:function(U){return},modelHydration:function(U){return}},QU={nextSibling:(U)=>{if(U.nextSibling)switch(U.nextSibling.type){case 0:return S(U.nextSibling);default:return U.nextSibling.domRef}return null},createTextNode:(U)=>{return document.createTextNode(U)},createElementNS:(U,z)=>{return document.createElementNS(U,z)},appendChild:(U,z)=>{return U.appendChild(z)},replaceChild:(U,z,G)=>{return U.replaceChild(z,G)},removeChild:(U,z)=>{return U.removeChild(z)},createElement:(U)=>{return document.createElement(U)},addClass:(U,z)=>{if(U)z.classList.add(U)},removeClass:(U,z)=>{if(U)z.classList.remove(U)},insertBefore:(U,z,G)=>{return U.insertBefore(z,G)},swapDOMRefs:(U,z,G)=>{let K=U.nextSibling;G.insertBefore(U,z),G.insertBefore(z,K);return},setInlineStyle:(U,z,G)=>{var K;for(let H in U)if(K=z[H],!K)if(H in G.style)G.style[H]="";else G.style.setProperty(H,"");else if(K!==U[H])if(H in G.style)G.style[H]=K;else G.style.setProperty(H,K);for(let H in z){if(U&&U[H])continue;if(H in G.style)G.style[H]=z[H];else G.style.setProperty(H,z[H])}return},setAttribute:(U,z,G)=>{return U.setAttribute(z,G)},setAttributeNS:(U,z,G,K)=>{return U.setAttributeNS(z,G,K)},removeAttribute:(U,z)=>{return U.removeAttribute(z)},setTextContent:(U,z)=>{U.textContent=z;return},flush:()=>{return},getHead:function(){return document.head},getRoot:function(){return document.body}};globalThis.miso={hydrationContext:HU,eventContext:s,drawingContext:QU,componentContext:KU,diff:A,hydrate:d,version:F,callBlur:I,callFocus:L,callSelect:w,callSetSelectionRange:V,eventJSON:B,fetchCore:O,eventSourceConnect:b,eventSourceClose:T,websocketConnect:P,websocketClose:f,websocketSend:j,updateRef:h,inline:m,typeOf:g,mathRandom:v,getRandomValues:u,splitmix32:y,populateClass:k,integrityCheck:a,delegateEvent:D,delegator:s.delegator,setDrawingContext:function(U){let z=globalThis[U].drawingContext,G=globalThis[U].eventContext;if(!z)console.error('Custom rendering engine ("drawingContext") is not defined at globalThis[name].drawingContext',U);if(!G)console.error('Custom event delegation ("eventContext") is not defined at globalThis[name].eventContext',U);globalThis.miso.drawingContext=z,globalThis.miso.eventContext=G}};
+var F="1.9.0.0";function L(U,z){var G=function(){var K=document.getElementById(U);if(K&&K.focus)K.focus()};z>0?setTimeout(G,z):G()}function I(U,z){var G=function(){var K=document.getElementById(U);if(K&&K.blur)K.blur()};z>0?setTimeout(G,z):G()}function w(U,z){var G=function(){var K=document.getElementById(U);if(K&&typeof K.select==="function")K.select()};z>0?setTimeout(G,z):G()}function V(U,z,G,K){var H=function(){var X=document.getElementById(U);if(X&&typeof X.setSelectionRange==="function")X.setSelectionRange(z,G,"none")};K>0?setTimeout(H,K):H()}function O(U,z,G,K,H,X,Q){var Z={method:z,headers:K};if(G)Z.body=G;let _={},$=null;try{fetch(U,Z).then((Y)=>{$=Y.status;for(let[W,J]of Y.headers)_[W]=J;if(!Y.ok)throw new Error(Y.statusText);if(Q=="json")return Y.json();else if(Q=="text")return Y.text();else if(Q==="arrayBuffer")return Y.arrayBuffer();else if(Q==="blob")return Y.blob();else if(Q==="bytes")return Y.bytes();else if(Q==="formData")return Y.formData();else if(Q==="none")return H({error:null,body:null,headers:_,status:$})}).then((Y)=>H({error:null,body:Y,headers:_,status:$})).catch((Y)=>X({error:null,body:Y,headers:_,status:$}))}catch(Y){X({body:null,error:Y.message,headers:_,status:$})}}function P(U,z,G,K,H,X,Q,Z,_){try{let $=new WebSocket(U);return $.onopen=function(){z()},$.onclose=function(Y){G(Y)},$.onerror=function(Y){console.error(Y),Z("WebSocket error received")},$.onmessage=function(Y){if(typeof Y.data==="string")try{if(_){if(K)K(Y.data);return}let W=JSON.parse(Y.data);if(H)H(W)}catch(W){if(_&&K)K(Y.data);else Z(W.message)}else if(Y.data instanceof Blob){if(X)X(Y.data)}else if(Y.data instanceof ArrayBuffer){if(Q)Q(Y.data)}else console.error("Received unknown message type from WebSocket",Y),Z("Unknown message received from WebSocket")},$}catch($){Z($.message)}}function f(U){if(U)U.close(),U=null}function j(U,z){if(z&&U&&U.readyState===WebSocket.OPEN)U.send(z)}function b(U,z,G,K,H,X){try{let Q=new EventSource(U);return Q.onopen=function(){z()},Q.onerror=function(){H("EventSource error received")},Q.onmessage=function(Z){try{if(X){if(G)G(Z.data);return}let _=JSON.parse(Z.data);if(K)K(_)}catch(_){if(X&&G)G(Z.data);else H(_.message)}},Q}catch(Q){H(Q.message)}}function T(U){if(U)U.close(),U=null}function k(U,z){if(!U.classList)U.classList=new Set;for(let G of z)for(let K of G.trim().split(" "))if(K)U.classList.add(K)}function h(U,z){if(!U.parent)return;z.nextSibling=U.nextSibling?null:U.nextSibling,z.parent=U.parent,U.parent.child=z}function g(U,z={}){let G=Object.keys(z),K=Object.values(z);return new Function(...G,U)(...K)}function m(U){if(U===null||U===void 0)return 0;if(typeof U==="number")return 1;if(typeof U==="string")return 2;if(typeof U==="boolean")return 3;if(Array.isArray(U))return 4;return 5}function y(U){return function(){U|=0,U=U+2654435769|0;var z=U^U>>>15;return z=Math.imul(z,2246822507),z=z^z>>>13,z=Math.imul(z,3266489909),((z^z>>>16)>>>0)/4294967296}}function u(){let U=new Uint32Array(1);return crypto.getRandomValues(U)[0]}function v(){return Math.random()}function q(U){switch(U.type){case 0:return S(U);default:return U.domRef}}function S(U){if(!U.child)throw new Error("'drill' called on an unmounted Component. This should never happen, please make an issue.");switch(U.child.type){case 0:return S(U.child);default:return U.child.domRef}}function A(U,z,G,K){if(!U&&!z)return;else if(!U)BU(z,G,K);else if(!z)c(U,G,K);else if(U.type===2&&z.type===2)XU(U,z,K);else if(U.type===0&&z.type===0){if(z.key===U.key){if(z.child=U.child,z.componentId=U.componentId,U.child)U.child.parent=z;return}l(U,z,G,K)}else if(U.type===1&&z.type===1)if(z.tag===U.tag&&z.key===U.key)z.domRef=U.domRef,o(U,z,K);else l(U,z,G,K);else l(U,z,G,K)}function XU(U,z,G){if(U.text!==z.text)G.setTextContent(U.domRef,z.text);z.domRef=U.domRef;return}function l(U,z,G,K){switch(U.type){case 2:break;default:M(U);break}switch(i(G,1,q(U),z,K),U.type){case 2:break;default:R(U);break}}function c(U,z,G){switch(U.type){case 2:break;default:M(U);break}switch(G.removeChild(z,q(U)),U.type){case 2:break;default:R(U);break}}function R(U){switch(YU(U),U.type){case 1:for(let z of U.children)if(z.type===1||z.type===0)R(z);break;case 0:if(U.child){if(U.child.type===1||U.child.type===0)R(U.child)}break}}function YU(U){if(U.type===1&&U.onDestroyed)U.onDestroyed();if(U.type===0)DU(U)}function ZU(U){switch(U.type){case 0:break;case 1:if(U.onBeforeDestroyed)U.onBeforeDestroyed();break;default:break}}function M(U){switch(ZU(U),U.type){case 1:for(let z of U.children){if(z.type===2)continue;M(z)}break;case 0:if(U.child){if(U.child.type===1||U.child.type===0)M(U.child)}break}}function o(U,z,G){_U(U?U.props:{},z.props,z.domRef,z.ns==="svg",G),$U(U?U.classList:null,z.classList,z.domRef,G),WU(U?U.css:{},z.css,z.domRef,G),AU(U?U.children:[],z.children,z.domRef,G),EU(z)}function $U(U,z,G,K){if(!U&&!z)return;if(!U){for(let H of z)K.addClass(H,G);return}if(!z){for(let H of U)K.removeClass(H,G);return}for(let H of U)if(!z.has(H))K.removeClass(H,G);for(let H of z)if(!U.has(H))K.addClass(H,G);return}function _U(U,z,G,K,H){var X;for(let Q in U)if(X=z[Q],X===void 0)if(K||!(Q in G)||Q==="disabled")H.removeAttribute(G,Q);else H.setAttribute(G,Q,"");else{if(X===U[Q]&&Q!=="checked"&&Q!=="value")continue;if(K)if(Q==="href")H.setAttributeNS(G,"http://www.w3.org/1999/xlink","href",X);else H.setAttribute(G,Q,X);else if(Q in G&&!(Q==="list"||Q==="form"))G[Q]=X;else H.setAttribute(G,Q,X)}for(let Q in z){if(U&&U[Q])continue;if(X=z[Q],K)if(Q==="href")H.setAttributeNS(G,"http://www.w3.org/1999/xlink","href",X);else H.setAttribute(G,Q,X);else if(Q in G&&!(Q==="list"||Q==="form"))G[Q]=z[Q];else H.setAttribute(G,Q,X)}}function WU(U,z,G,K){K.setInlineStyle(U,z,G)}function qU(U,z){if(U.length===0||z.length===0)return!1;for(var G=0;G<U.length;G++)if(U[G].key===null||U[G].key===void 0)return!1;for(var G=0;G<z.length;G++)if(z[G].key===null||z[G].key===void 0)return!1;return!0}function AU(U,z,G,K){if(qU(U,z))RU(U,z,G,K);else for(let H=0;H<Math.max(z.length,U.length);H++)A(U[H],z[H],G,K)}function JU(U,z){if(U.ns==="svg")U.domRef=z.createElementNS("http://www.w3.org/2000/svg",U.tag);else if(U.ns==="mathml")U.domRef=z.createElementNS("http://www.w3.org/1998/Math/MathML",U.tag);else U.domRef=z.createElement(U.tag)}function n(U,z,G){if(z.onCreated)z.onCreated(z.domRef)}function i(U,z,G,K,H){switch(K.type){case 2:switch(K.domRef=H.createTextNode(K.text),z){case 2:H.insertBefore(U,K.domRef,G);break;case 0:H.appendChild(U,K.domRef);break;case 1:H.replaceChild(U,K.domRef,G);break}break;case 0:SU(U,z,G,K,H);break;case 1:if(K.onBeforeCreated)K.onBeforeCreated();if(JU(K,H),K.onCreated)K.onCreated(K.domRef);switch(o(null,K,H),z){case 2:H.insertBefore(U,K.domRef,G);break;case 0:H.appendChild(U,K.domRef);break;case 1:H.replaceChild(U,K.domRef,G);break}break}}function EU(U){if(U.tag==="canvas"&&U.draw)U.draw(U.domRef)}function DU(U){U.unmount(U.componentId)}function SU(U,z,G,K,H){let X=K.mount(U);if(K.componentId=X.componentId,K.child=X.componentTree,X.componentTree.parent=K,X.componentTree.type!==0){let Q=q(X.componentTree);if(z===1&&G)H.replaceChild(U,Q,G);else if(z===2)H.insertBefore(U,Q,G)}}function BU(U,z,G){i(z,0,null,U,G)}function N(U,z,G,K){K.insertBefore(U,q(z),G?q(G):null)}function NU(U,z,G,K){K.swapDOMRefs(q(U),q(z),G)}function RU(U,z,G,K){var H=0,X=0,Q=U.length-1,Z=z.length-1,_,$,Y,W,J,C,r;for(;;){if(X>Z&&H>Q)break;if($=z[X],Y=z[Z],J=U[H],W=U[Q],H>Q)A(null,$,G,K),N(G,$,J,K),U.splice(X,0,$),X++;else if(X>Z){_=Q;while(Q>=H)c(U[Q--],G,K);U.splice(H,_-H+1);break}else if(J.key===$.key)A(U[H++],z[X++],G,K);else if(W.key===Y.key)A(U[Q--],z[Z--],G,K);else if(J.key===Y.key&&$.key===W.key)NU(W,J,G,K),MU(U,H,Q),A(U[H++],z[X++],G,K),A(U[Q--],z[Z--],G,K);else if(J.key===Y.key)N(G,J,W.nextSibling,K),U.splice(Q,0,U.splice(H,1)[0]),A(U[Q--],z[Z--],G,K);else if(W.key===$.key)N(G,W,J,K),U.splice(H,0,U.splice(Q,1)[0]),A(U[H++],$,G,K),X++;else{C=!1,_=H;while(_<=Q){if(U[_].key===$.key){C=!0,r=U[_];break}_++}if(C)U.splice(H,0,U.splice(_,1)[0]),A(U[H++],$,G,K),N(G,r,U[H],K),X++;else i(G,2,q(J),$,K),U.splice(H++,0,$),X++,Q++}}}function MU(U,z,G){let K=U[z];U[z]=U[G],U[G]=K}function UU(U,z,G,K,H){for(let X of z)H.addEventListener(U,X.name,function(Q){CU(Q,U,G,K,H)},X.capture)}function CU(U,z,G,K,H){G(function(X){if(Array.isArray(U))for(let Q of U)t(Q,X,z,K,H);else t(U,X,z,K,H)})}function t(U,z,G,K,H){var X=H.getTarget(U);if(X){let Q=FU(G,X,H);D(U,z,Q,K,H)}}function FU(U,z,G){var K=[];while(!G.isEqual(U,z))if(K.unshift(z),z&&G.parentNode(z))z=G.parentNode(z);else return K;return K}function D(U,z,G,K,H){if(!G.length){if(K)console.warn('Event "'+U.type+'" did not find an event handler to dispatch on',z,U);return}else if(G.length>1){if(z.type===2)return;else if(z.type===0){if(!z.child){if(K)throw console.error("VComp has no child property set during event delegation",z),console.error("This means the Component has not been fully mounted, this should never happen"),new Error("VComp has no .child property set during event delegation");return}return D(U,z.child,G,K,H)}else if(z.type===1){if(H.isEqual(z.domRef,G[0])){let X=z.events.captures[U.type];if(X){let Q=X.options;if(Q.preventDefault)U.preventDefault();if(!U.captureStopped)X.runEvent(U,z.domRef);if(Q.stopPropagation)U.captureStopped=!0}G.splice(0,1);for(let Q of z.children)if(H.isEqual(q(Q),G[0]))D(U,Q,G,K,H)}return}}else if(z.type===0){if(z.child)D(U,z.child,G,K,H)}else if(z.type===1){let X=z.events.captures[U.type];if(X&&!U.captureStopped){let Z=X.options;if(H.isEqual(G[0],z.domRef)){if(Z.preventDefault)U.preventDefault();if(X.runEvent(U,G[0]),Z.stopPropagation)U.captureStopped=!0}}let Q=z.events.bubbles[U.type];if(Q&&!U.captureStopped){let Z=Q.options;if(H.isEqual(G[0],z.domRef)){if(Z.preventDefault)U.preventDefault();if(Q.runEvent(U,G[0]),!Z.stopPropagation)e(z.parent,U)}}else if(!U.captureStopped)e(z.parent,U)}}function e(U,z){while(U)switch(U.type){case 2:break;case 1:let G=U.events.bubbles[z.type];if(G){let K=G.options;if(K.preventDefault)z.preventDefault();if(G.runEvent(z,U.domRef),K.stopPropagation)return}U=U.parent;break;case 0:if(!U.eventPropagation)return;U=U.parent;break}}function B(U,z){if(typeof U[0]==="object"){var G=[];for(var K=0;K<U.length;K++)G.push(B(U[K],z));return G}for(let Z of U)z=z[Z];var H;if(z instanceof Array||"length"in z&&z.localName!=="select"){H=[];for(var X=0;X<z.length;X++)H.push(B([],z[X]));return H}H={};for(var Q in LU(z)){if(z.localName==="input"&&(Q==="selectionDirection"||Q==="selectionStart"||Q==="selectionEnd"))continue;if(typeof z[Q]=="string"||typeof z[Q]=="number"||typeof z[Q]=="boolean")H[Q]=z[Q]}return H}function LU(U){var z={},G=0;do{var K=Object.getOwnPropertyNames(U);for(G=0;G<K.length;G++)z[K[G]]=null}while(U=Object.getPrototypeOf(U));return z}function IU(U){var z=0,G=U.length>0?[U[0]]:[];for(var K=1;K<U.length;K++){if(G[z].type===2&&U[K].type===2){G[z].text+=U[K].text;continue}G[++z]=U[K]}return G}function d(U,z,G,K,H){if(!G||!z)return!1;if(z.nodeType===3)return!1;if(!p(U,G,K.firstChild(z),K,H)){if(U)console.warn("[DEBUG_HYDRATE] Could not copy DOM into virtual DOM, falling back to diff");while(K.firstChild(z))H.removeChild(z,K.lastChild(z));return!1}else if(U)console.info("[DEBUG_HYDRATE] Successfully prerendered page");return!0}function x(U,z,G){if(U)console.warn("[DEBUG_HYDRATE] VTree differed from node",z,G)}function zU(U){if(U.substr(0,1)=="#"){let z=(U.length-1)/3,G=[17,1,0.062272][z-1];return[Math.round(parseInt(U.substr(1,z),16)*G),Math.round(parseInt(U.substr(1+z,z),16)*G),Math.round(parseInt(U.substr(1+2*z,z),16)*G)]}else return U.split("(")[1].split(")")[0].split(",").map((z)=>{return+z})}function a(U,z,G){return GU(!0,U,z,G)}function GU(U,z,G,K){if(z.type==2){if(G.getTag(z.domRef)!=="#text")console.warn("VText domRef not a TEXT_NODE",z),U=!1;else if(z.text!==G.getTextContent(z.domRef))console.warn("VText node content differs",z),U=!1}else if(z.type===1){if(z.tag.toUpperCase()!==G.getTag(z.domRef).toUpperCase())console.warn("Integrity check failed, tags differ",z.tag.toUpperCase(),G.getTag(z.domRef)),U=!1;if("children"in z&&z.children.length!==G.children(z.domRef).length)console.warn("Integrity check failed, children lengths differ",z,z.children,G.children(z.domRef)),U=!1;for(let H in z.props)if(H==="href"||H==="src"){let X=window.location.origin+"/"+z.props[H],Q=G.getAttribute(z.domRef,H),Z=z.props[H];if(X!==Q&&Z!==Q&&Z+"/"!==Q&&X+"/"!==Q)console.warn("Property "+H+" differs",z.props[H],G.getAttribute(z.domRef,H)),U=!1}else if(H==="height"||H==="width"){if(parseFloat(z.props[H])!==parseFloat(G.getAttribute(z.domRef,H)))console.warn("Property "+H+" differs",z.props[H],G.getAttribute(z.domRef,H)),U=!1}else if(H==="class"||H==="className"){if(z.props[H]!==G.getAttribute(z.domRef,"class"))console.warn("Property class differs",z.props[H],G.getAttribute(z.domRef,"class")),U=!1}else if(z.props[H]!==G.getAttribute(z.domRef,H))console.warn("Property "+H+" differs",z.props[H],G.getAttribute(z.domRef,H)),U=!1;for(let H in z.css)if(H==="color"){if(zU(G.getInlineStyle(z.domRef,H)).toString()!==zU(z.css[H]).toString())console.warn("Style "+H+" differs",z.css[H],G.getInlineStyle(z.domRef,H)),U=!1}else if(z.css[H]!==G.getInlineStyle(z.domRef,H))console.warn("Style "+H+" differs",z.css[H],G.getInlineStyle(z.domRef,H)),U=!1;for(let H of z.children){let X=GU(U,H,G,K);U=U&&X}}return U}function p(U,z,G,K,H){switch(z.type){case 0:if(!p(U,z.child,G,K,H))return!1;break;case 2:if(G.nodeType!==3||z.text.trim()!==G.textContent.trim())return x(U,z,G),!1;z.domRef=G;break;case 1:if(G.nodeType!==1)return x(U,z,G),!1;z.domRef=G,z.children=IU(z.children),n(G,z,H);for(var X=0;X<z.children.length;X++){let Q=z.children[X],Z=G.childNodes[X];if(!Z)return x(U,Q,Z),!1;if(!p(U,Q,Z,K,H))return!1}break}return!0}var s={addEventListener:(U,z,G,K)=>{U.addEventListener(z,G,K)},delegator:(U,z,G,K,H)=>{UU(U,z,G,K,H)},isEqual:(U,z)=>{return U===z},getTarget:(U)=>{return U.target},parentNode:(U)=>{return U.parentNode}},HU={getInlineStyle:(U,z)=>{return U.style[z]},firstChild:(U)=>{return U.firstChild},lastChild:(U)=>{return U.lastChild},getAttribute:(U,z)=>{if(z==="class")return U.className;if(z in U)return U[z];return U.getAttribute(z)},getTag:(U)=>{return U.nodeName},getTextContent:(U)=>{return U.textContent},children:(U)=>{return U.childNodes}},KU={mountComponent:function(U,z,G){return},unmountComponent:function(U){return},modelHydration:function(U){return}},QU={nextSibling:(U)=>{if(U.nextSibling)switch(U.nextSibling.type){case 0:return S(U.nextSibling);default:return U.nextSibling.domRef}return null},createTextNode:(U)=>{return document.createTextNode(U)},createElementNS:(U,z)=>{return document.createElementNS(U,z)},appendChild:(U,z)=>{return U.appendChild(z)},replaceChild:(U,z,G)=>{return U.replaceChild(z,G)},removeChild:(U,z)=>{return U.removeChild(z)},createElement:(U)=>{return document.createElement(U)},addClass:(U,z)=>{if(U)z.classList.add(U)},removeClass:(U,z)=>{if(U)z.classList.remove(U)},insertBefore:(U,z,G)=>{return U.insertBefore(z,G)},swapDOMRefs:(U,z,G)=>{let K=U.nextSibling;G.insertBefore(U,z),G.insertBefore(z,K);return},setInlineStyle:(U,z,G)=>{var K;for(let H in U)if(K=z[H],!K)if(H in G.style)G.style[H]="";else G.style.setProperty(H,"");else if(K!==U[H])if(H in G.style)G.style[H]=K;else G.style.setProperty(H,K);for(let H in z){if(U&&U[H])continue;if(H in G.style)G.style[H]=z[H];else G.style.setProperty(H,z[H])}return},setAttribute:(U,z,G)=>{return U.setAttribute(z,G)},setAttributeNS:(U,z,G,K)=>{return U.setAttributeNS(z,G,K)},removeAttribute:(U,z)=>{return U.removeAttribute(z)},setTextContent:(U,z)=>{U.textContent=z;return},flush:()=>{return},getHead:function(){return document.head},getRoot:function(){return document.body}};globalThis.miso={hydrationContext:HU,eventContext:s,drawingContext:QU,componentContext:KU,diff:A,hydrate:d,version:F,callBlur:I,callFocus:L,callSelect:w,callSetSelectionRange:V,eventJSON:B,fetchCore:O,eventSourceConnect:b,eventSourceClose:T,websocketConnect:P,websocketClose:f,websocketSend:j,updateRef:h,inline:g,typeOf:m,mathRandom:v,getRandomValues:u,splitmix32:y,populateClass:k,integrityCheck:a,delegateEvent:D,delegator:s.delegator,setDrawingContext:function(U){let z=globalThis[U].drawingContext,G=globalThis[U].eventContext;if(!z)console.error('Custom rendering engine ("drawingContext") is not defined at globalThis[name].drawingContext',U);if(!G)console.error('Custom event delegation ("eventContext") is not defined at globalThis[name].eventContext',U);globalThis.miso.drawingContext=z,globalThis.miso.eventContext=G}};
