+++ ts/miso.ts
@@ -42,36 +42,34 @@ import {
     HydrationContext,
     NodeId,
     VTreeType,
-    ComponentId
+    ComponentId,
+    Component,
+    Runtime,
+    PATCH,
+    InsertBefore,
+    SwapDOMRefs,
+    CreateElement,
+    CreateElementNS,
+    CreateTextNode,
+    SetAttribute,
+    SetAttributeNS,
+    AppendChild,
+    RemoveChild,
+    ReplaceChild,
+    RemoveAttribute,
+    SetTextContent,
+    SetInlineStyle,
+    MountComponent,
+    UnmountComponent,
+    ModelHydration,
+    AddClass,
+    RemoveClass,
+    ProcessEvent,
+    AddEventListeners,
+    Flush
 } from './miso/types';
 
-import {
-  patch,
-  PATCH,
-  Component,
-  Runtime,
-  InsertBefore,
-  SwapDOMRefs,
-  CreateElement,
-  CreateElementNS,
-  CreateTextNode,
-  SetAttribute,
-  SetAttributeNS,
-  AppendChild,
-  RemoveChild,
-  ReplaceChild,
-  RemoveAttribute,
-  SetTextContent,
-  SetInlineStyle,
-  MountComponent,
-  UnmountComponent,
-  ModelHydration,
-  AddClass,
-  RemoveClass,
-  ProcessEvent,
-  AddEventListeners,
-  Flush,
-} from './miso/patch';
+import { patch } from './miso/patch';
 
 import { patchDrawingContext } from './miso/context/patch';
 import { vcomp, vnode, vtext } from './miso/smart';
